Trying to fix DELETE
