+++ src/app/api/senior/route.schema.ts
@@ -1,0 +1,18 @@
+import { z } from "zod";
+import { unauthorizedErrorSchema, unknownErrorSchema } from "../route.schema";
+
+export const postSeniorSchema = z.object({
+  name: z.string(),
+  location: z.string(),
+  description: z.string(),
+  StudentIDs: z.array(z.string()),
+  ChapterID: z.string(),
+});
+
+export type IPostSeniorRequestSchema = z.infer<typeof postSeniorSchema>;
+
+export const seniorPostResponse = z.discriminatedUnion("code", [
+  z.object({ code: z.literal("SUCCESS") }),
+  unknownErrorSchema,
+  unauthorizedErrorSchema,
+]);
