+++ src/app/private/[uid]/admin/elist/test/page.tsx
@@ -3,18 +3,18 @@
 import {
   deleteSenior,
   patchSenior,
-  postSenior,
 } from "src/app/api/senior/[id]/route.client";
-import { IPostSeniorRequestSchema } from "@api/senior/[id]/route.schema";
+import { postSenior } from "@api/senior/route.client";
+import { IPostSeniorRequestSchema } from "@api/senior/route.schema";
 
 function Test() {
   async function Poo() {
     postSenior({
-      userId: "6587829665af0d81089c42fb",
+      // userId: "6587829665af0d81089c42fb",
       body: {
-        name: "Stephen",
-        location: "Boston",
-        description: "He is cool",
+        name: "Nathan",
+        location: "Nevada",
+        description: "Cool beans",
         StudentIDs: ["6587829665af0d81089c42fb"],
         ChapterID: "65878595b94284e0c3e02d55",
       },
@@ -24,24 +24,29 @@ function Test() {
   }
 
   async function Delete() {
-    deleteSenior({
-      seniorId: "65bfe4fe217927ba65687658",
-      body: {
-        userId: "6587829665af0d81089c42fb",
-      },
-    });
+    deleteSenior({ seniorId: "65c4263cdcb4f869ef2f2ba9" });
   }
 
-  // async function Patch() {
-  //   patchSenior({
-
-  //   })
-  // }
+  async function Patch() {
+    patchSenior({
+      seniorId: "65c4263cdcb4f869ef2f2ba9",
+      body: {
+        name: "Nathan",
+        location: "Arkansas",
+        description: "Cool beans",
+        StudentIDs: ["6587829665af0d81089c42fb"],
+        ChapterID: "65878595b94284e0c3e02d55",
+      },
+    }).then((res) => {
+      console.log(res);
+    });
+  }
 
   return (
     <div>
       <button onClick={Poo}>Add senior</button>
-      <button onClick={Delete}>Delete</button>
+      <button onClick={Patch}>Patch this</button>
+      <button onClick={Delete}>Delete senior</button>
     </div>
   );
 }
